+++ packages/eas-cli/src/__tests__/commands/workflow-runs-test.ts
@@ -6,7 +6,7 @@ import {
   mockTestCommand,
 } from './utils';
 import WorkflowRunList from '../../commands/workflow/runs';
-import { WorkflowRunStatus } from '../../graphql/generated';
+import { WorkflowRunStatus, WorkflowRunTriggerEventType } from '../../graphql/generated';
 import { AppQuery } from '../../graphql/queries/AppQuery';
 import { WorkflowRunQuery } from '../../graphql/queries/WorkflowRunQuery';
 import { enableJsonOutput, printJsonOnlyOutput } from '../../utils/json';
@@ -75,8 +75,10 @@ describe(WorkflowRunList, () => {
       {
         id: 'build',
         status: WorkflowRunStatus.Success,
+        triggerEventType: WorkflowRunTriggerEventType.Manual,
         createdAt: '2022-01-01T00:00:00.000Z',
         updatedAt: '2022-01-01T00:00:00.000Z',
+        errors: [],
         gitCommitHash: '1234567890',
         gitCommitMessage: 'commit message',
         workflow: {
@@ -119,6 +121,8 @@ describe(WorkflowRunList, () => {
         status: 'FAILURE',
         gitCommitHash: '1234567890',
         gitCommitMessage: 'commit message',
+        trigger: '',
+        triggerType: 'Other',
         startedAt: '2022-01-01T00:00:00.000Z',
         finishedAt: '2022-01-01T00:00:00.000Z',
         workflowId: 'build',
@@ -135,10 +139,12 @@ describe(WorkflowRunList, () => {
       {
         id: 'build1',
         status: WorkflowRunStatus.Failure,
+        triggerEventType: WorkflowRunTriggerEventType.Manual,
         createdAt: '2022-01-01T00:00:00.000Z',
         updatedAt: '2022-01-01T00:00:00.000Z',
         gitCommitHash: '1234567890',
         gitCommitMessage: 'commit message',
+        errors: [],
         workflow: {
           id: 'build',
           name: 'build',
@@ -165,6 +171,8 @@ describe(WorkflowRunList, () => {
         status: 'FAILURE',
         gitCommitHash: '1234567890',
         gitCommitMessage: 'commit message',
+        trigger: '',
+        triggerType: 'Manual',
         startedAt: '2022-01-01T00:00:00.000Z',
         finishedAt: '2022-01-01T00:00:00.000Z',
         workflowId: 'build',